Prefer "designation" over "name" for 1PIF fields
For a WebForm field, the "name" value is the HTML field name on the
target site and the "designation" value is an enum value that indicates
known data values that have specific handling in 1Password (primarily
"username" and "password").  Because the "name" value varies from site
to site, the mapping to the known importer fields is not consistently
recognized, which results in imported entries that are missing the
"login" or "password" values.  The "designation" value provides a key
name that is consistent across all sites, so it should be preferred if
it is available.