Don't import "trashed" 1PIF entries
Newer versions of 1Password retain deleted items (including items that
are moved to a different vault) until explicitly purged by the user,
and these deleted items are included in the 1PIF export.  The deleted
items will include a "trashed" field with a value of `true`, and this
can be used to filter these items out of the import.

A recent update to 1Password has converted the "Trash" container, which
held _all_ deleted entries, into an "Archive" container which allows
items to be explicitly archived if they might need to be restored later.
This allows items which are truly no longer needed to be intentionally
deleted, and those items will be entirely removed from the vault and
will not be included in the 1PIF export.  However, it is still unlikely
that a user would want the 1Password "Archive" entries to be imported
into "pass" as active items.

A better long-term solution for archived items might be to import them
and then immediately delete them so that the item is preserved in the
password store Git history, but this is not a feature that is currently
supported by the shared importer infrastructure.

See https://blog.1password.com/introducing-archive/