Ignore some additional 1PIF fields
Newer versions of 1Password have added additional fields to the 1PIF
export that can clutter the output if the `--all` option is used.  These
fields don't provide any useful information, and should be ignored.

Note that the "sections" field contains all custom data added to a
1Password entry, and therefore could contain useful information, but
the format of the data makes it effectively useless without additional
processing.